requirements.txt
python-dotenv
aiogram
google-genai
sqlmodel
apscheduler
Pillow
requests
